chore: bump project version to 2.0.0
I've decided to properly follow semantic versioning. The old automarker latest release was 1.2.3. This is a complete rework which looks different and is used differently, so it makes sense to increase the major version number. Afer this, the minor version will be increased when any new features are added (with these features not forcing anyone to completely re-learn the app). The patch version will be increased when there are any bug fixes/refactorings.